+++ src/FsCodec/StringId.fs
@@ -1,0 +1,20 @@
+namespace FsCodec
+
+/// Endows any type that inherits this class with standard .NET comparison semantics using a supplied token identifier
+[<AbstractClass>]
+type Comparable<'TComp, 'Token when 'TComp :> Comparable<'TComp, 'Token> and 'Token: comparison>(token: 'Token) =
+    member private _.Token = token
+    override x.Equals y = match y with :? Comparable<'TComp, 'Token> as y -> x.Token = y.Token | _ -> false
+    override _.GetHashCode() = hash token
+    interface System.IComparable with
+        member x.CompareTo y =
+            match y with
+            | :? Comparable<'TComp, 'Token> as y -> compare x.Token y.Token
+            | _ -> invalidArg "y" "invalid comparand"
+
+/// Endows any type that inherits this class with standard .NET comparison semantics using a supplied token identifier
+/// + treats the token as the canonical rendition for `ToString()` purposes
+[<AbstractClass>]
+type StringId<'TComp when 'TComp :> Comparable<'TComp, string>>(token: string) =
+    inherit Comparable<'TComp, string>(token)
+    override _.ToString() = token